Restructure files for packaging
