--- pylana/pylana_v2.py
@@ -1,7 +1,0 @@
-from pylana.modules.logs import LogsAPI
-from pylana.modules.resources import ResourceAPI
-from pylana.modules.shiny_dashboards import ShinyDashboardAPI
-
-
-class LanaAPI2(LogsAPI, ShinyDashboardAPI, ResourceAPI):
-    pass
